Refactor: Split MissionStats component
Refactors MissionStats.tsx into smaller, more manageable components and hooks, improving code organization and maintainability.
